Fix portability issues related to Decklink SDK and OSX
Minor fixups to allow building on OSX.  In particular, allow
the user to point to a standalone SDK since the version committed
into this repository only compiles on Linux.  Also adjust the
string related functions to accommodate the fact that on platforms
other than Linux the strings are not a "char *".