Make klvanc_capture properly parse standard definition VANC lines
When in an SDI video mode that is standard definition, the VANC
data spans both the Luma and Chroma channels.  Make the parser
properly do the colorspace conversion so that VANC can be found
in standard definition video feeds.

We'll probably later make an API change to fold this functionality
into libklvanc itself, but this gets the job done for now.

Note that this change is to the display parser only, not the
recording functionality (which required no changes).  Hence the
tool should continue work against VANC capture files previously
recorded with klvanc_capture (and files previously captured
against SD feeds will now properly decode and display the VANC).