Fix case where we were using wrong CSC routine on HD video
The caller for file processing was passing the stride instead of
the pixel width, which was inconsistent with how live capture
works.  This caused the wrong CSC to be used for live data display,
as well having entries appear three times since the parser was
overflowing the buffer and reading theree times as much data.

Make the file processor pass the width just like live capture,
and fix the check which determines whether the video is HD or SD
to reference the width instead of the stride.