Add support for video mode autodetection
If the user doesn't specify "-m <mode>", autodetect the video mode
from the source signal (and handle changes in mode after capture
has been started).

Note, if the mode argument is specified, capture will be locked to
that mode and will *not* change if the input source doesn't match
at startup or changes over time.