Don't display every frame of audio by default when using -A
When using the "-A" argument to split the audio into PCM files for
each pair, by default it also dumps all the audio data to the display.

Require the user to specify "-v" to get this behavior.

Prior to this change, splitting 1 minute of audio would take 26
seconds.  With the change it takes 0.830 seconds.