Fix use of condition variable on main thread
The pthread_cond_wait() man page makes clear that there can be
spurious wakeups and that the function should always be used in
conjunction with a boolean check for the global it protects.

Make sure we set the global *before* waking the condition variable,
and check the variable after the wait so it gets restarted if there
is a spurious wakeup.