Get rid of need to seek backwards when reading muxed streams
The existing logic relied on the first member of every frame
header type to be some constant value, and then the demux
would read four bytes, get the constant, seek backwards four
bytes, and then call the appropriate function to decode the
frame.

This approach results in not being able to work with non-seekable
sources such as sockets/named pipes, as well as not being
practical to use with compressed content (since rewinding on
a gzip stream requires the entire stream to be decompressed again).

Move the type indicator outside of the header struct for each frame
type.  This allows the type to be read to be read without having
to rewind before calling the individual frame decoder.

Note that nothing in the underlying file format is being changed
by this patch - this is entirely an artifact of how the file is
parsed and generated, unrelated to what the resulting file looks
like on disk.