Move -std=c++11 to CXXFLAGS
The C++ standard directive is not a preprocessor directive, and
thus needs to be in CXX flags rather than the CPP flags.

Having it in the CPPFLAGS causes it to be passed to both the C and
C++ compilers, which results as a warning for gcc and and under
Clang's C++ compiler it causes a compile error.