Make sure C++ binaries compiled with "-g" when ./configure --enable-debug
Do some cleanup of the CFLAGS/CXXFLAGS to ensure that C++ binaries
get compiled with "-g" when enabling debug in ./configure.  This
patch also ensures that C++ programes get compiled with "-Wall"

Note: the addition of this patch causes warnings at compile-time to
be output that were not visible before.  Separate patches can
address the warnings now being reported.