Add minimal support for burning frame counter into video
Add a minimalist routine to burn arbitrary text into v210 frames,
and use it to burn a frame counter into the video.  Also, insert
a line when a VANC packet is inserted, as well as a count since
the last VANC packet was injected.  This will aid in validating
SCTE-104 timing for downstream consumers.