Cleanup warnings when -Wall is enabled
Take a pass over the warnings reported by the compiler when built
with -Wall.

Note: I've commented out the lastSilenceIdx declaration rather than
deleting it because it's used by a commented out printf() line, but
it seems like the debug line wouldn't work even when uncommented since
nothing in the code sets lastSilenceIdx.