Move the code that burns status info until after the VANC is processed
Move the code block further down, until after the VANC is processed.
The primary motiviation for this is the counter since last insert
was saying the first frame after the frame with VANC was zero.  Moving
the block ensures that the frame with the VANC is "0 frames since last
insert", and then the counter starts at one on the next frame.

(i.e. the output was confusing to view after the fact in analyzers
since it was off by one).