klvanc_transmitter: fix case where input file is not v210 aligned
The klvanc_transmitter was taking advantage of the fact that
klvanc_capture creates VANC sample files which are v210 aligned.
If the files aren't aligned on a six-pixel boundary the output will
be truncated.  In any case, the utility was putting out a bunch of
blank padding after VANC packets, which was less than ideal.

Rework the logic to use the utility function already provided by
the klvanc library to generate the v210 buffer from the array of
16-bit Y samples.  This utility function properly handles inputs
which aren't aligned on the six-pixel boundary, and properly pads
out as needed with the correct values.