Add support for burning KL frame counters into video as well
In addition to inserting KL frame counters into VANC, also burn
them into the video in a manner parseable by the burnreader.  This
allows klvanc_capture to check the synchrnonization between video
and VANC.

When specifying the "-k" option for klvanc_capture in curses mode,
an extra line will be inserted into the GUI which shows how far
out of sync the video and VANC are.