Don't spew video/vanc lines unless "-k" is specified
Before this patch I was using the "-k" to only control whether
the synchronization info is shown in the ncurses GUI.  Tweak
the logic so we don't check the synchronization at all unless
"-k" is specified.  This avoids spewing out one line per video
frame to the console unless the user specifically wants to see
the synchronization info.