klvanc_transmitter: Make VANC output work in SD modes
If doing 480i/576i video, the VANC packets need to span both
the luma and chroma pixels (as opposed to HD where they only
go into the luma region).  Tweak the logic so that we properly
insert the VANC packets for SD resolutions.

This was causing klvanc_transmitter to inject malformed packets
that couldn't be parsed by downstream receivers if selecting the
NTSC or PAL output video format.