Make burncounters work in SD resolutions
The burncounters work fine for 720p and higher, but for standard
definition the boxes are too large, which prevents them from
all fitting on one line.

Change the logic such that in HD modes the counters are 30x30 pixels
but on SD resolutions they are 18x18.  This lets them fit onto a
single line without blowing past the end of the line.

Note that the check for SD vs. HD is a bit hacky as the v210
functions only receive the horizontal stride rather than the pixel
width of the line.  Hence I picked a value for the stride that
works for both transmitter and capture (which for non-obvious
reasons the two tools use different strides for a given resolution).
This logic can be improved in the future if needed.