Adjust which line we read to parse the burnreader counters
Some devices may slightly shift the video vertically, and cause
the burnreader to not find the counters because it was reading on
the very first line.

Adjust the logic to read on line 10, which is about 1/3 of the
way through for HD and halfway down for SD.  This make the reader
a bit more robust by ensuring that slight reformatting of the
picture still results in the counters being found.

The problem was detected in a commercial SDI video processor, where
it shifted the video down by a single line, and that was enough to
cause the reader to not be able to find the burnreader codes.